Add `test` function
