+++ crates/forge/src/run_tests/workspace.rs
@@ -1,26 +1,16 @@
 use super::package::RunForPackageArgs;
-use crate::profile_validation::check_profile_compatibility;
 use crate::run_tests::messages::latest_blocks_numbers::LatestBlocksNumbersMessage;
 use crate::run_tests::messages::overall_summary::OverallSummaryMessage;
 use crate::run_tests::messages::tests_failure_summary::TestsFailureSummaryMessage;
-use crate::warn::{
-    error_if_snforge_std_deprecated_missing, error_if_snforge_std_deprecated_not_compatible,
-    error_if_snforge_std_not_compatible, warn_if_backtrace_without_panic_hint,
-    warn_if_snforge_std_deprecated_does_not_match_package_version,
-};
 use crate::{
-    ColorOption, ExitStatus, MINIMAL_SCARB_VERSION_FOR_V2_MACROS_REQUIREMENT, TestArgs,
-    block_number_map::BlockNumberMap, run_tests::package::run_for_package,
+    ExitStatus, TestArgs, block_number_map::BlockNumberMap, run_tests::package::run_for_package,
     scarb::build_artifacts_with_scarb, shared_cache::FailedTestsCache,
-    warn::warn_if_snforge_std_does_not_match_package_version,
 };
 use anyhow::{Context, Result};
 use forge_runner::test_case_summary::AnyTestCaseSummary;
 use forge_runner::{CACHE_DIR, test_target_summary::TestTargetSummary};
 use foundry_ui::UI;
-use scarb_api::metadata::{MetadataOpts, metadata_with_opts};
 use scarb_api::{
-    ScarbCommand,
     metadata::{Metadata, PackageMetadata},
     target_dir_for_workspace,
 };
@@ -30,41 +20,18 @@ use std::env;
 use std::sync::Arc;
 
 #[tracing::instrument(skip_all, level = "debug")]
-pub async fn run_for_workspace(args: TestArgs, ui: Arc<UI>) -> Result<ExitStatus> {
-    match args.color {
-        // SAFETY: This runs in a single-threaded environment.
-        ColorOption::Always => unsafe { env::set_var("CLICOLOR_FORCE", "1") },
-        // SAFETY: This runs in a single-threaded environment.
-        ColorOption::Never => unsafe { env::set_var("CLICOLOR", "0") },
-        ColorOption::Auto => (),
-    }
-
-    let scarb_metadata = metadata_with_opts(MetadataOpts {
-        profile: args.scarb_args.profile.specified(),
-        ..MetadataOpts::default()
-    })?;
-
-    check_profile_compatibility(&args, &scarb_metadata)?;
-
-    let scarb_version = ScarbCommand::version().run()?.scarb;
-    if scarb_version >= MINIMAL_SCARB_VERSION_FOR_V2_MACROS_REQUIREMENT {
-        error_if_snforge_std_not_compatible(&scarb_metadata)?;
-        warn_if_snforge_std_does_not_match_package_version(&scarb_metadata, &ui)?;
-    } else {
-        error_if_snforge_std_deprecated_missing(&scarb_metadata)?;
-        error_if_snforge_std_deprecated_not_compatible(&scarb_metadata)?;
-        warn_if_snforge_std_deprecated_does_not_match_package_version(&scarb_metadata, &ui)?;
-    }
-
-    warn_if_backtrace_without_panic_hint(&scarb_metadata, &ui);
-
+pub async fn run_for_workspace(
+    scarb_metadata: &Metadata,
+    args: TestArgs,
+    ui: Arc<UI>,
+) -> Result<ExitStatus> {
     let artifacts_dir_path =
         target_dir_for_workspace(&scarb_metadata).join(&scarb_metadata.current_profile);
 
     let packages: Vec<PackageMetadata> = args
         .scarb_args
         .packages_filter
-        .match_many(&scarb_metadata)
+        .match_many(scarb_metadata)
         .context("Failed to find any packages matching the specified filter")?;
 
     let filter = PackagesFilter::generate_for::<Metadata>(packages.iter());
